fix/LAB1-6: Исправлена структура проекта, код стал более читаемый
