+++ backend/swagger.js
@@ -1,7 +1,6 @@
 // swagger.js
 const swaggerJsDoc = require('swagger-jsdoc');
 const swaggerUi = require('swagger-ui-express');
-
 const swaggerOptions = {
     swaggerDefinition: {
         openapi: '3.0.0',
@@ -20,5 +19,4 @@ const swaggerOptions = {
 };
 
 const swaggerDocs = swaggerJsDoc(swaggerOptions);
-
 module.exports = { swaggerUi, swaggerDocs };
